[#16] Move data source schema to data_source file. Move terraform models from HTTP Client package to internal service package (tags). Replace built-in golang types with terraform framework types
