--- internal/service/tags/schema.go
@@ -1,38 +1,0 @@
-package tags
-
-import (
-	"context"
-
-	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
-)
-
-func (d *TagsDataSource) Schema(ctx context.Context, req datasource.SchemaRequest, resp *datasource.SchemaResponse) {
-	resp.Schema = schema.Schema{
-		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: "Use this resource to retrieve the tags associated with the Statsig Project.",
-
-		Attributes: map[string]schema.Attribute{
-			"tags": schema.ListNestedAttribute{
-				Computed: true,
-				NestedObject: schema.NestedAttributeObject{
-					Attributes: map[string]schema.Attribute{
-						// id, name, description, isCore
-						"id": schema.StringAttribute{
-							Computed: true,
-						},
-						"name": schema.StringAttribute{
-							Computed: true,
-						},
-						"description": schema.StringAttribute{
-							Computed: true,
-						},
-						"is_core": schema.BoolAttribute{
-							Computed: true,
-						},
-					},
-				},
-			},
-		},
-	}
-}
